refactor:폴더구조 컨벤션맞추어 수정 #78
